index: add support for "void elements"
@Coreh this breaks your previous "childless nodes" test
(adjusted to be fixed), so please take a look at this
and make sure nothing on Editor is breaking afterwards.

Will probably just do a major version bump after this
since it's a rather big change.

Unrelated: we should probably convert this to TypeScript
or ES6 to benefit from some compile-time checking.